refactor(OrderTable at el): after separate packages, move entity, service, etc about OrderTable
