+++ src/main/java/kitchenpos/table/dao/OrderTableDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.table.dao;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 import java.util.List;
 import java.util.Optional;
