+++ src/test/java/kitchenpos/tablegroup/ui/TableGroupRestControllerTest.java
@@ -2,9 +2,8 @@ package kitchenpos.tablegroup.ui;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import kitchenpos.tablegroup.application.TableGroupService;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
-import kitchenpos.tablegroup.ui.TableGroupRestController;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
